requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.29
pdfplumber==0.11.0
openai==1.30.0
python-multipart==0.0.9
pydantic==2.7.1
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
python-dotenv==1.0.1
